+++ src/units/AngularVelocity.cpp
@@ -8,7 +8,7 @@ AngularVelocity::AngularVelocity() : m_radiansPerSecond(0) {
 AngularVelocity::~AngularVelocity() {
 }
 
-float AngularVelocity::getRadiansPerSecond() const {
+double AngularVelocity::getRadiansPerSecond() const {
     return m_radiansPerSecond;
 }
