+++ src/units/Centimeters.h
@@ -7,7 +7,7 @@ namespace sim {
 class Centimeters : public Distance {
 
 public:
-    Centimeters(float centimeters);
+    Centimeters(double centimeters);
     Centimeters(const Distance& distance);
     Centimeters operator+(const Distance& distance) const;
