+++ src/units/Duration.cpp
@@ -8,15 +8,15 @@ Duration::Duration() : m_seconds(0) {
 Duration::~Duration() {
 }
 
-float Duration::getSeconds() const {
+double Duration::getSeconds() const {
     return m_seconds;
 }
 
-float Duration::getMilliseconds() const {
+double Duration::getMilliseconds() const {
     return getSeconds() * 1000;
 }
 
-float Duration::getMicroseconds() const {
+double Duration::getMicroseconds() const {
     return getSeconds() * 1000 * 1000;
 }
