+++ src/units/Meters.h
@@ -9,13 +9,13 @@ class Meters : public Distance {
 
 public:
     Meters();
-    Meters(float meters);
+    Meters(double meters);
     Meters(const Distance& distance);
     Meters operator+(const Distance& distance) const;
     Meters operator-(const Distance& distance) const;
-    Meters operator*(float factor) const;
-    Meters operator/(float factor) const;
-    float operator/(const Distance& distance) const;
+    Meters operator*(double factor) const;
+    Meters operator/(double factor) const;
+    double operator/(const Distance& distance) const;
     MetersSquared operator*(const Distance& distance) const;
     void operator+=(const Distance& distance);
