+++ src/units/Seconds.h
@@ -7,7 +7,7 @@ namespace sim {
 class Seconds : public Duration {
 
 public:
-    Seconds(float seconds);
+    Seconds(double seconds);
     Seconds(const Duration& duration);
     Seconds operator+(const Duration& duration) const;
