+++ src/units/Speed.cpp
@@ -8,7 +8,7 @@ Speed::Speed() : m_metersPerSecond(0) {
 Speed::~Speed() {
 }
 
-float Speed::getMetersPerSecond() const {
+double Speed::getMetersPerSecond() const {
     return m_metersPerSecond;
 }
